Fix the ordering of grub-install
Puts grub-install of the create-live-image.sh script
before the system partitions are unmounted.
Previously, the installation step happened when the work/mnt
directory had the config partition mounted, instead of the root filesystem.

Adds --root-directory flag to grub-install, making GRUB
scan only the loopback device's partition. Previously, with
this option missing, host system was scanned and the LiveUSB
would not work if built on one system and booted on a different one.

Both of those changes make it possible to actually boot the image,
tested using QEMU.